+++ bot/commands/info/link.js
@@ -2,10 +2,14 @@
 var mediaStore = require(process.cwd()+ '/bot/store/mediaInfo.js');
 var _ = require('lodash');
 
+function makeYTCheckerUrl(yid){
+  return `https://polsy.org.uk/stuff/ytrestrict.cgi?ytid=${yid}`;
+}
+
 module.exports = function(bot, db, data) {
   if(!data) { return; }
-  var current = mediaStore.getCurrent();
-
+  var current = mediaStore.getCurrent();  
+  var restrictions = '';
 
   var whoAsked = _.get(data, 'user.username', '');
   if (whoAsked !== ''){ 
@@ -13,12 +17,11 @@ module.exports = function(bot, db, data) {
   }
 
   if(!current.link) {
-      bot.sendChat('No song is playing at this time!');
-  } else{
-      if (!current.link) {
-        bot.sendChat('Sorry I couldn\'t get the link for this track for some reason');
-      } else {
-        bot.sendChat(`${whoAsked} The current song is ${current.name}, and the link is ${current.link}`);
-      }
+    bot.sendChat('Sorry I couldn\'t get the link for this track');
+  } else {
+    if (current.type === 'youtube') {
+      restrictions = `YouTube region restriction info: ${makeYTCheckerUrl(current.id)}`;
+    }
+    bot.sendChat(`${whoAsked} - ${current.link} - ${restrictions}`);
   }
 };